Re-register users to getUserSubs schedule at server reboot
